Use host configuration for URL [#166]
