+++ src/main/kotlin/com/hexagonkt/todokt/backend/entities/Task.kt
@@ -1,0 +1,12 @@
+package com.hexagonkt.todokt.backend.entities
+
+import com.hexagonkt.settings.SettingsManager
+
+data class Task(
+    val id: String,
+    val title: String,
+    val order: Int? = null,
+    val completed: Boolean? = null
+){
+    val url = "${SettingsManager.requireSetting("serviceUrl") as String}/tasks/$id"
+}
